requirement.txt
python == 3.6
pytorch == 1.5.0